requirements.txt
Django >=4.2, <5.0
djangorestframework >= 3.15.1
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.27.1
pillow>=10.3.0
django-cors-headers
razorpay
